screenshare/requirements.txt
accelerate==0.16.0
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.2
bidict==0.22.1
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
click==8.0.1
colorama==0.4.4
comtypes==1.1.10
decorator==4.4.2
Desktopmagic==14.3.11
diffusers==0.12.1
Django==4.1.5
djangorestframework==3.14.0
exceptiongroup==1.1.0
filelock==3.9.0
Flask==2.0.1
frozenlist==1.3.3
gTTS==2.2.3
h11==0.14.0
html5lib==1.1
huggingface-hub==0.12.0
idna==2.10
imageio==2.25.0
imageio-ffmpeg==0.4.8
importlib-metadata==6.0.0
itsdangerous==2.0.1
Jinja2==3.0.1
keyboard==0.13.5
MarkupSafe==2.0.1
MouseInfo==0.1.3
moviepy==1.0.3
mpg123==0.4
mss==7.0.1
multidict==6.0.4
mysql-connector-python==8.0.25
numpy==1.24.2
openai==0.26.5
outcome==1.2.0
packaging==23.0
Pillow==9.4.0
proglog==0.1.10
protobuf==3.17.3
psutil==5.9.4
pyasn1==0.4.8
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.6
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
pytesseract==0.3.10
python-engineio==4.3.4
python-socketio==5.7.2
pyttsx3==2.90
pytweening==1.0.4
pytz==2022.7.1
pywin32==301
pywinauto==0.6.8
PyYAML==6.0
regex==2022.10.31
requests==2.25.1
rsa==4.9
selenium==4.8.0
sentencepiece==0.1.97
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tokenizers==0.13.2
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.64.1
transformers==4.26.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.5
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.0.1
wget==3.2
WMI==1.5.1
wsproto==1.2.0
yarl==1.8.2
zipp==3.13.0
